HCO_arduino_only2.py is backwards compatable with half_center5.pde and half_center4.pde Both sketches are confirmed working.

+++ serial/half_center5/half_center5.pde
@@ -94,7 +94,7 @@ class NeuronGroup { public:
      } 
   }
   
-  void spiking(int neuron, int var, int weight) {
+  void spiking(int neuron, int var, double weight) {
     neurons[neuron].values[var] += weight;
   }
